basic flow done

+++ TaskBuilderObjC/Modules/Components/Task/TaskDelegate.h
@@ -11,8 +11,8 @@
 
 @protocol TaskDelegate
 
-- (void) progress: (NSString*) taskName_ progress:(double)value_;
-- (void) stateChanged: (NSString*) taskName_ progress:(double)value_;
+- (void) progress: (NSString*) taskName_ progress:(float)value_;
+- (void) stateChanged: (NSString*) taskName_ progress:(float)value_;
 
 //- (void)progress:(Task)task;
 //- (void)stateChanged:(RowData *)task;
